fix intake constants file
