+++ packages/core/src/domain/configuration/configuration.ts
@@ -12,9 +12,9 @@ import { TrackingConsent } from '../trackingConsent'
 import type { SessionPersistence } from '../session/sessionConstants'
 import type { MatchOption } from '../../tools/matchOption'
 import { isAllowedTrackingOrigins } from '../allowedTrackingOrigins'
+import type { Site } from '../../tools/intakeConstants'
 import type { TransportConfiguration } from './transportConfiguration'
 import { computeTransportConfiguration } from './transportConfiguration'
-import type { Site } from './intakeSites'
 
 export const DefaultPrivacyLevel = {
   ALLOW: 'allow',
