+++ packages/core/src/domain/configuration/index.ts
@@ -8,5 +8,4 @@ export {
 } from './configuration'
 export type { EndpointBuilder, TrackType } from './endpointBuilder'
 export { createEndpointBuilder, buildEndpointHost } from './endpointBuilder'
-export * from './intakeSites'
 export { computeTransportConfiguration, isIntakeUrl } from './transportConfiguration'
