+++ packages/core/src/domain/telemetry/telemetry.spec.ts
@@ -3,8 +3,7 @@ import { callMonitored } from '../../tools/monitor'
 import type { ExperimentalFeature } from '../../tools/experimentalFeatures'
 import { resetExperimentalFeatures, addExperimentalFeatures } from '../../tools/experimentalFeatures'
 import type { Configuration } from '../configuration'
-import { INTAKE_SITE_US1_FED } from '../../tools/intakeConstants'
-import { INTAKE_SITE_US1 } from '../configuration'
+import { INTAKE_SITE_US1_FED, INTAKE_SITE_US1 } from '../../tools/intakeConstants'
 import { setNavigatorOnLine, setNavigatorConnection } from '../../../test'
 import type { Context } from '../../tools/serialisation/context'
 import { Observable } from '../../tools/observable'
