+++ packages/core/src/tools/intakeConstants.ts
@@ -10,4 +10,11 @@ export type Site =
   | 'ap1.datadoghq.com'
 
 export const INTAKE_SITE_STAGING: Site = 'datad0g.com' as Site
+export const INTAKE_SITE_FED_STAGING: Site = 'dd0g-gov.com' as Site
+export const INTAKE_SITE_US1: Site = 'datadoghq.com'
+export const INTAKE_SITE_EU1: Site = 'datadoghq.eu'
 export const INTAKE_SITE_US1_FED: Site = 'ddog-gov.com'
+
+export const PCI_INTAKE_HOST_US1 = 'pci.browser-intake-datadoghq.com'
+export const INTAKE_URL_PARAMETERS = ['ddsource', 'ddtags']
+
